Sort Kubernetes object before installing them.
There are certain interdependencies between Kubernetes objects that
need to be taken care of, e.g. a Namespace must be created before any
object using that namespace, or a CustomResourceDefinition needs to
exist before any manifest that is of the kind defined by CRDs.

The list is heavily inspired by Helm's kind_sorter.go at:

https://github.com/helm/helm/blob/8c84a0bc0376650bc3d7334eef0c46356c22fa36/pkg/releaseutil/kind_sorter.go

Signed-off-by: Milan Plzik <milan.plzik@grafana.com>